split class in three
