+++ core/src/main/java/com/novoda/noplayer/internal/exoplayer/error/ErrorFormatter.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.internal.exoplayer.forwarder;
+package com.novoda.noplayer.internal.exoplayer.error;
 
 final class ErrorFormatter {
