+++ core/src/main/java/com/novoda/noplayer/internal/exoplayer/error/ExoPlayerErrorMapper.java
@@ -1,0 +1,31 @@
+package com.novoda.noplayer.internal.exoplayer.error;
+
+import com.google.android.exoplayer2.ExoPlaybackException;
+import com.novoda.noplayer.DetailErrorType;
+import com.novoda.noplayer.NoPlayer;
+import com.novoda.noplayer.NoPlayerError;
+import com.novoda.noplayer.PlayerErrorType;
+
+import static com.novoda.noplayer.internal.exoplayer.error.ErrorFormatter.formatMessage;
+
+public final class ExoPlayerErrorMapper {
+
+    private ExoPlayerErrorMapper() {
+        // Static class.
+    }
+
+    public static NoPlayer.PlayerError errorFor(ExoPlaybackException exception) {
+        String message = formatMessage(exception.getCause());
+
+        switch (exception.type) {
+            case ExoPlaybackException.TYPE_SOURCE:
+                return SourceErrorMapper.map(exception.getSourceException(), message);
+            case ExoPlaybackException.TYPE_RENDERER:
+                return RendererErrorMapper.map(exception.getRendererException(), message);
+            case ExoPlaybackException.TYPE_UNEXPECTED:
+                return UnexpectedErrorMapper.map(exception.getUnexpectedException(), message);
+            default:
+                return new NoPlayerError(PlayerErrorType.UNKNOWN, DetailErrorType.UNKNOWN, formatMessage(exception.getCause()));
+        }
+    }
+}
